% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpur.R
\name{gpu_demo}
\alias{gpu_demo}
\title{DEPRECATED -- gpuR no longer available on CRAN
Demos power of GPU at matrix multiplication}
\usage{
gpu_demo()
}
\description{
DEPRECATED -- gpuR no longer available on CRAN
Demos power of GPU at matrix multiplication
}
