% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_interactive.R
\name{interactive_similarity}
\alias{interactive_similarity}
\title{compstatslib interactive_similarity() function}
\usage{
interactive_similarity(points = data.frame(), ...)
}
\description{
compstatslib interactive_similarity() function
}
