% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{machine_precision}
\alias{machine_precision}
\title{compstatslib machine_precision() function}
\usage{
machine_precision()
}
\value{
A numeric value representing the smallest possible number that the user's computer can effectively represent.
}
\description{
Code function that reports the smallest possible number on the user's machine such that 1 + x != 1.
}
\examples{
machine_precision()

# Anything smaller is swallowed by rounding
1 + machine_precision() != 1
1 + machine_precision() / 2 != 1

}
