% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{plot_regr}
\alias{plot_regr}
\title{compstatslib plot_regr(points) function}
\usage{
plot_regr(points, regression = TRUE, stats = TRUE)
}
\description{
Non-interactive visualization function that plots given points, regression line and relevant statistics.
}
\details{
params:
  points - dataframe of x,y points to plot
}
