% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_statistic_plot.R
\name{plot_t_test}
\alias{plot_t_test}
\title{compstatslib plot_t_test() function}
\usage{
plot_t_test(diff = 0.5, sd = 4, n = 100, alpha = 0.05, error_matrix = FALSE)
}
\arguments{
\item{diff}{The test difference (defaults to 0.5).}

\item{sd}{Population standard deviation (defaults to 4).}

\item{n}{Sample size (defaults to 100).}

\item{alpha}{Significance level (defaults to 0.05).}

\item{error_matrix}{Whether to display type I/II error matrix (defaults to FALSE).}
}
\value{
No return value, called for side effects (plots to the active
device).
}
\description{
Non-interactive visualization function that plots null and alternative
t distributions of a t-test. Shows rejection zone and power as areas
under the curves.
}
\examples{
plot_t_test()
plot_t_test(diff=-0.1, sd=3)

}
\seealso{
\code{\link{interactive_t_test}}
}
