% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_interactive.R
\name{interactive_logit}
\alias{interactive_logit}
\title{compstatslib interactive_logit() function}
\usage{
interactive_logit(
  points = data.frame(),
  formula = y ~ x,
  min_x = 0,
  max_x = 50,
  ...
)
}
\arguments{
\item{points}{An optional \code{dataframe} of x and y points to plot
and estimate the regression.
If no \code{points} are provided, the user is free to click and create points
on the plot area.}

\item{formula}{A \code{formula} to use in estimating logit
(e.g., \code{y ~ x}).}

\item{min_x}{Minimum x value for the plot range.}

\item{max_x}{Maximum x value for the plot range.}

\item{...}{Further arguments passed to the \code{plot_logit()} function that
produces the plot.}
}
\value{
A \code{dataframe} containing the points coordinates.
}
\description{
Interactive visualization function that lets you point-and-click to add
data points, while it automatically plots and updates a logistic regression
curve and associated statistics.
}
\details{
Click on the plotting area to add points and see a corresponding logistic
regression curve. Click "Done" to return the points to the console.
}
\examples{
# Selecting coordinates on the plot area, storing them in 'pts'
pts <- interactive_logit()

# Replotting the points stored earlier in 'pts', allowing the user to
# continue the interactive regression
interactive_logit(pts)

}
\seealso{
\code{\link{plot_logit}}
}
