% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_inverse_interactive.R
\name{interactive_matrix_inverse}
\alias{interactive_matrix_inverse}
\title{compstatslib interactive_matrix_inverse() function}
\usage{
interactive_matrix_inverse(x1_init = 1, y1_init = 2, x2_init = 2, y2_init = 1)
}
\arguments{
\item{x1_init}{The first row (or column) vector of matrix A. This parameter is set to 1 by default.}

\item{y1_init}{The second row (or column) vector of matrix A. This parameter is set to 2 by default.

The area of the parallelogram resulting from these two vectors is the determinant of matrix A.}

\item{x2_init}{The first row (or column) vector of the inverse matrix A^(-1). This parameter is set to 2 by default.}

\item{y2_init}{The second row (or column) vector of the inverse matrix A^(-1). This parameter is set to 1 by default.

The area of the parallelogram resulting from these two vectors is the determinant of the inverse matrix A^(-1).}
}
\description{
Interactive function that allows one to visualize a matrix inversion
using sliders to adjust matrix parameters.
}
\details{
Inspired by: \link{https://math.stackexchange.com/questions/295250/geometric-interpretations-of-matrix-inverses/1922830#1922830?newreg=d55776428dd2418e8a6a971dfdbfa17}


Use the sliders in the viewer to adjust the matrix parameters and see the
resulting transformation. Click "Done" to close.
}
\seealso{
\code{\link{plot_matrix_inverse}}
}
