% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_inverse.R
\name{interactive_matrix_inverse}
\alias{interactive_matrix_inverse}
\title{Interactive demonstration of matrix determinants
Inspired by: https://math.stackexchange.com/questions/295250/geometric-interpretations-of-matrix-inverses/1922830#1922830?newreg=d55776428dd2418e8a6a971dfdbfa17}
\usage{
interactive_matrix_inverse(x1_init = 1, y1_init = 2, x2_init = 2, y2_init = 1)
}
\description{
Interactive demonstration of matrix determinants
Inspired by: https://math.stackexchange.com/questions/295250/geometric-interpretations-of-matrix-inverses/1922830#1922830?newreg=d55776428dd2418e8a6a971dfdbfa17
}
