% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_interactive.R
\name{interactive_pca}
\alias{interactive_pca}
\title{compstatslib interactive_pca() function}
\usage{
interactive_pca(meancenter = TRUE)
}
\arguments{
\item{meancenter}{A logical parameter that will mean center the points if
set to TRUE, while it will not mean center the points if set to FALSE. By
default it is set to TRUE.}
}
\value{
A list containing:
 \item{points}{A dataframe with \code{x} and \code{y} coordinates.}
 \item{pca}{The \code{prcomp} result if 3+ points were added, or
 \code{NULL} otherwise.}
}
\description{
Interactive visualization function that lets you point-and-click to add data
points, while it automatically plots and updates principal component vectors.
}
\details{
Click on the plotting area to add points and see corresponding principal
components. Click "Done" to return results to the console.
}
\seealso{
\code{\link{plot_pca}}
}
