% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_interactive.R
\name{interactive_pca}
\alias{interactive_pca}
\title{compstatslib interactive_pca() function}
\usage{
interactive_pca()

Click on the plotting area to add points and see corresponding principal components (hitting ESC will stop the simulation).
}
\arguments{
\item{meancenter}{A logical parameter that will mean center the points if set to TRUE, while it will not mean center the points if set to FALSE. By default it is set to TRUE.}
}
\value{
A dataframe containing the *x* and *y* coordinates of the points selected by the user, and a list of parameters related to the estimated principal components, including:
 \item{sdev}{A vector of the standard deviations of the estimated principal components.}
 \item{rotation}{A dataframe of the principal components coordinates.}
 \item{center}{A vector of length equal the number of columns of x can be supplied. The value is passed to *scale*.}
 \item{scale}{A logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place. The default is FALSE.}
 \item{x}{A numeric or complex matrix (or data frame) which provides the data for the principal components analysis.}
}
\description{
Interactive visualization function that lets you point-and-click to add data points, while it automatically plots and updates principal component vectors.
}
