% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_interactive.R
\name{interactive_regression}
\alias{interactive_regression}
\title{interactive_regression() runs a regression simulation.
Click on the plotting area to add points and see a corresponding regression line
 (hitting ESC will stop the simulation).}
\usage{
interactive_regression(points = data.frame(), ...)
}
\description{
You can supply the following parameter:
  old_pts - dataframe of x,y points to start interacting with
}
\details{
You will also see three numbers: 
  intercept – where the regression line intercepts the y-axis
  regression coefficient – the slope of x on y
  correlation - correlation of x and y
  r-squared - R^2 of y
}
