% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_statistic_interactive.R
\name{interactive_t_test}
\alias{interactive_t_test}
\title{compstatslib interactive_t_test() function}
\usage{
interactive_t_test()
}
\arguments{
\item{diff}{The test difference.}

\item{sd}{Population standard deviation.}

\item{n}{Sample size.}

\item{alpha}{Significance level.}
}
\description{
Interactive visualization that allows one to adjust the parameters affecting
hypothesis testing in order to see how their variation influences the null t
and alternative t distributions, and statistical power.
}
\details{
Use the sliders in the viewer to adjust parameters. The movement of the
alternative t-statistics distribution with respect to the null distribution
will be visible, as well as the consequent change in statistical power.
Click "Done" to close.
}
\seealso{
\code{\link{plot_t_test}}
}
