% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_statistic_interactive.R
\name{interactive_t_test}
\alias{interactive_t_test}
\title{compstatslib interactive_t_test() function}
\usage{
interactive_t_test()

One can click on the gear icon on the top-left corner of the plotting area to open the parameter settings.
The movement of the alternative t-statistics distribution with respect to the null distribution will be visible, as well as the consequent change in statistical power.
}
\arguments{
\item{diff}{The test difference.}

\item{sd}{Population standard deviation.}

\item{n}{Sample size.}

\item{alpha}{Significance level.}
}
\description{
Interactive visualization function that allows one to *manipulate* the parameters that affect hypothesis testing in order to see how their variation influences the null t and alternative t distributions, and statistical power.
}
\seealso{
\code{\link{plot_t_test}}
}
