% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{machine_precision}
\alias{machine_precision}
\title{Reports smallest possible number on machine such that 1 + x != 1}
\usage{
machine_precision()
}
\description{
Reports smallest possible number on machine such that 1 + x != 1
}
