% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_plot.R
\name{plot_logit}
\alias{plot_logit}
\title{compstatslib plot_logit() function}
\usage{
plot_logit(
  points,
  formula = y ~ x,
  regression = TRUE,
  stats = TRUE,
  min_x = 0,
  max_x = 50,
  legend_loc = "topleft"
)
}
\description{
compstatslib plot_logit() function
}
