% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_plot.R
\name{plot_logit}
\alias{plot_logit}
\title{compstatslib plot_logit() function}
\usage{
plot_regr(points, ...)
}
\arguments{
\item{points}{A \code{dataframe} of *x* and *y* coordinates to plot.}

\item{formula}{A \code{formula} to use in estimating logit
(e.g., \code{y ~ x}).}

\item{regression}{Logical parameter of whether to plot a regression line 
(TRUE by default)}

\item{stats}{Logical parameter of whether to display relevant statistics on
the plot area (TRUE by default).
The following parameters are provided on the plot area:
 intercept of the logistic regression;
 coefficient of the independent variable;
 AIC (Akaike information criterion).}
}
\value{
A \code{dataframe} containing the points coordinates.
}
\description{
Non-interactive visualization function that plots given points,
logistic regression line and relevant statistics.
}
\examples{
pts <- data.frame(x = c(1, 3, 5, 8), y = c(2, 4, 6 ,8))
plot_regr(pts, formula = y ~ x)

names(pts) <- c("input", "output")
plot_regr(pts, formula = output ~ input)

}
\seealso{
\code{\link{interactive_regression}}
}
