% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_plot.R
\name{plot_logit}
\alias{plot_logit}
\title{compstatslib plot_logit() function}
\usage{
plot_logit(points, ...)
}
\arguments{
\item{points}{A \code{dataframe} of *x* and *y* coordinates to plot.}

\item{formula}{A \code{formula} to use in estimating logit
(e.g., \code{y ~ x}).}

\item{regression}{Logical parameter of whether to plot a regression line
(TRUE by default)}

\item{stats}{Logical parameter of whether to display relevant statistics on
the plot area (TRUE by default).
The following parameters are provided on the plot area:
\enumerate{
 \item{Intercept}{The intercept of the logistic regression.}
 \item{Coefficient}{The coefficient of the independent variable.}
 \item{AIC}{An information theoretic fit criterion.}
}}
}
\value{
A \code{dataframe} containing the points coordinates.
}
\description{
Non-interactive visualization function that plots given points,
logistic regression line and relevant statistics.
}
\examples{
mydata <- data.frame(iv = c(-6, -3, 1, 3, 5, 8), dv = c(0, 0, 0, 1, 1 ,1))
plot_logit(mydata, formula = dv ~ iv)

}
\seealso{
\code{\link{interactive_regression}}
}
