% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{plot_regr}
\alias{plot_regr}
\title{#' plot_regr(points) plots given points, regression line, stats}
\usage{
plot_regr(points, regression = TRUE, stats = TRUE)
}
\description{
params:
  points - dataframe of x,y points to plot
}
