% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{plot_regr}
\alias{plot_regr}
\title{compstatslib plot_regr() function}
\usage{
plot_regr(points, ...)
}
\arguments{
\item{points}{A \code{dataframe} of *x* and *y* coordinates to plot.}

\item{regression}{A logical parameter that plots a regression line when set to TRUE and hides it when set to FALSE. It is set to TRUE by default.}

\item{stats}{A logical parameter that displays the relevant statistics on the plot area when set to TRUE; it hides them when set to FALSE. It is set to TRUE by default.}
}
\value{
A \code{dataframe} containing the points coordinates. Additionally, the following parameters are provided on the plot area:
 \item{Raw intercept}{The y-coordinate at which the regression line crosses the y-axis.}
 \item{Raw slope}{The value of the slope parameter.}
 \item{Correlation}{The strength of the linear relationship.}
 \item{SSR}{The sum of squares regression.}
 \item{SSE}{The sum of squares error.}
 \item{SST}{The sum of squares total.}
 \item{R-squared}{The multiple coefficient of determination.}
}
\description{
Non-interactive visualization function that plots given points, regression line and relevant statistics.
}
\examples{
points <- data.frame(x = c(1, 3, 5, 8), y = c(2, 4, 6 ,8))

plot_regr(points)

}
\seealso{
\code{\link{interactive_regression}}
}
