% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_plot.R
\name{plot_sampling}
\alias{plot_sampling}
\title{commpstatslib plot_sampling() function}
\usage{
plot_sampling()
}
\arguments{
\item{population}{A \code{vector} of values following any population distribution you wish to simulate.}

\item{sample_size}{The size of each sample to draw from the population}

\item{theta}{The \code{function} that computes the statistic of interest from each sample (e.g., \code{mean} or \code{median}).}
}
\value{
A \code{list} containing the population data, sample statistics, and other information of interest
}
\description{
Plot distribution of a population, samples drawn from the population, given sampling statistic.
}
\examples{
\dontrun{
plot_sampling(rnorm(100000), sample_size = 100, reps = 50, theta = median)
}

}
